fix(sqlgen): narrow integral EAV operands through the write funnel's float64 for DOUBLE-width classes (#384)
Fourth-review P1: with the DOUBLE projection, an integral operand above
2^53 still bound as exact int64 on the Postgres EAV routes while DuckDB's
CAST(? AS DOUBLE) rounded it — so equals:9007199254740993 missed a stored
float64 image 9007199254740992 on hot Postgres and matched it on every
DuckDB tier. NarrowEAVNumericOperand converts integral operands for
integer/smallint/numeric EAV comparisons to float64 — the same narrowing
the stored data took — in both the predicate normalizer and the batch
attr-value anchor (which must reach the same verdict on the same literal,
#355). bigint keeps exact int64 binds: its storage is BIGINT on every
tier. Docs §6.4 wording tightened; 2^53-boundary probes added to the
acceptance e2e on all three tiers.

+++ internal/sqlgen/sql_generator_test.go
@@ -73,10 +73,11 @@ func TestSQLGenerator_ToSQLClauses(t *testing.T) {
 		t.Fatalf("unexpected SQL clause.\nexpected: %s\nactual:   %s", expectedClause, sqlClause)
 	}
 
-	// price is `gt:10` on a numeric attribute: integral literals bind as exact
-	// int64 since #281 (fractional literals still bind float64).
+	// price is `gt:10` on a numeric attribute: integral literals narrow to
+	// the write funnel's float64 for the DOUBLE-width classes (#384;
+	// bigint keeps exact int64 per #281).
 	expectedArgs := []any{
-		int16(10), int64(10),
+		int16(10), float64(10),
 		int16(11), "active",
 		int16(12), "A%",
 	}
